requirements.txt
joblib==0.14.1
numpy==1.14.0
pandas==0.25.3
Pillow==7.2.0
pip==20.3.4
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.22.2.post1
scipy==1.4.1
setuptools==50.3.2
six==1.15.0
torch==1.2.0
torchvision==0.4.0
wheel==0.36.2